render/requirements.txt
python-telegram-bot==13.15
apscheduler==3.6.3
python-dotenv==1.0.1
requests==2.31.0
Flask==2.3.2
